chore(server): move blueprint
